backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
qdrant-client==1.7.0
openai==1.3.5
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
tiktoken==0.5.1